feat: minor refactors
