+++ aea/cli/generate_all_protocols.py
@@ -39,69 +39,40 @@ import shutil
 import subprocess  # nosec
 import sys
 import tempfile
-from io import StringIO
-from itertools import chain
-from operator import methodcaller
 from pathlib import Path
-from typing import Any, Iterator, List, Match, Optional, Tuple, cast
+from typing import Any, List, Match, cast
 
 import click
 import semver
 
 from aea.cli.registry.utils import download_file, extract, request_api
 from aea.common import JSONLike
-from aea.configurations.base import ComponentType, ProtocolConfig, ProtocolSpecification
-from aea.configurations.constants import DEFAULT_PROTOCOL_CONFIG_FILE
+from aea.configurations.base import ComponentType, ProtocolConfig
+from aea.configurations.constants import (
+    DEFAULT_PROTOCOL_CONFIG_FILE,
+    DEFAULT_README_FILE,
+    PROTOCOLS,
+)
 from aea.configurations.data_types import PackageId, PublicId
 from aea.configurations.loader import (
-    ConfigLoader,
     ConfigLoaders,
     load_component_configuration,
+    load_protocol_specification_from_string,
 )
 from aea.exceptions import enforce
+from aea.helpers.git import check_working_tree_is_dirty
+from aea.manager.project import AEAProject
 
 
 SPECIFICATION_REGEX = re.compile(r"(---\nname.*\.\.\.)", re.DOTALL)
 LIBPROTOC_VERSION = "libprotoc 3.19.4"
 CUSTOM_TYPE_MODULE_NAME = "custom_types.py"
-README_FILENAME = "README.md"
-PROTOCOLS_PLURALS = "protocols"
 PROTOCOL_GENERATOR_DOCSTRING_REGEX = "It was created with protocol buffer compiler version `libprotoc .*` and aea version `.*`."
-DEFAULT_TEST_SAMPLES = [
-    ("sample_specification.yaml", "t_protocol",),
-    ("sample_specification_no_custom_types.yaml", "t_protocol_no_ct",),
-]
 
 
 logging.basicConfig(format="[%(asctime)s][%(levelname)s] %(message)s")
 logger = logging.getLogger("generate_all_protocols")
 logger.setLevel(logging.INFO)
-
-
-def check_working_tree_is_dirty() -> None:
-    """Check if the current Git working tree is dirty."""
-    click.echo("Checking whether the Git working tree is dirty...")
-    result = subprocess.check_output(  # nosec
-        [str(shutil.which("git")), "diff", "--stat"]
-    )  # nosec
-    if len(result) > 0:
-        click.echo("Git working tree is dirty:")
-        click.echo(result.decode("utf-8"))
-        sys.exit(1)
-    else:
-        click.echo("All good!")
-
-
-def load_protocol_specification_from_string(
-    specification_content: str,
-) -> ProtocolSpecification:
-    """Load a protocol specification from string."""
-    file = StringIO(initial_value=specification_content)
-    config_loader = ConfigLoader(
-        "protocol-specification_schema.json", ProtocolSpecification
-    )
-    protocol_spec = config_loader.load_protocol_specification(file)
-    return protocol_spec
 
 
 def get_protocol_specification_from_readme(package_path: Path) -> str:
@@ -125,18 +96,12 @@ def get_protocol_specification_from_readme(package_path: Path) -> str:
     return specification_content
 
 
-def subdirs(path: Path) -> Iterator[Path]:
-    """Get subdirectories of a path."""
-    return filter(methodcaller("is_dir"), path.iterdir())
-
-
-def find_protocols_in_local_registry(packages_dir: Path) -> Iterator[Path]:
+def find_protocols_in_local_registry(packages_dir: Path) -> List[Path]:
     """Find all protocols in local registry."""
-    authors = subdirs(packages_dir)
-    component_parents = chain(*map(subdirs, authors))
-    protocols_parent = filter(lambda p: p.name == PROTOCOLS_PLURALS, component_parents)
-    protocols = chain(*map(subdirs, protocols_parent))
-    return map(methodcaller("absolute"), protocols)
+    return [
+        p.parent.absolute()
+        for p in packages_dir.glob(f"**/{DEFAULT_PROTOCOL_CONFIG_FILE}")
+    ]
 
 
 def log(message: str, level: int = logging.INFO) -> None:
@@ -147,57 +112,6 @@ def log(message: str, level: int = logging.INFO) -> None:
 def log_separator(length: int = 100) -> None:
     """Log separator."""
     click.echo("=" * length)
-
-
-def run_cli(*args: Any, **kwargs: Any) -> None:
-    """Run a CLI command."""
-    log(f"Calling command {args} with kwargs {kwargs}")
-    return_code = subprocess.check_call(args, **kwargs)  # nosec
-    enforce(
-        return_code == 0,
-        f"Return code of {pprint.pformat(args)} is {return_code} != 0.",
-    )
-
-
-def run_aea(*args: Any, **kwargs: Any) -> None:
-    """
-    Run an AEA command.
-
-    :param args: the AEA command
-    :param kwargs: keyword arguments to subprocess function
-    """
-    run_cli(sys.executable, "-m", "aea.cli", *args, **kwargs)
-
-
-class AEAProject:
-    """A context manager class to create and delete an AEA project."""
-
-    old_cwd: str
-    temp_dir: str
-
-    def __init__(self, name: str = "my_aea", parent_dir: Optional[str] = None):
-        """
-        Initialize an AEA project.
-
-        :param name: the name of the AEA project.
-        :param parent_dir: the parent directory.
-        """
-        self.name = name
-        self.parent_dir = parent_dir
-
-    def __enter__(self) -> None:
-        """Create and enter into the project."""
-        self.old_cwd = os.getcwd()
-        self.temp_dir = tempfile.mkdtemp(dir=self.parent_dir)
-        os.chdir(self.temp_dir)
-
-        run_aea("create", "--local", "--empty", self.name, "--author", "fetchai")
-        os.chdir(self.name)
-
-    def __exit__(self, exc_type, exc_val, exc_tb) -> None:  # type: ignore
-        """Exit the context manager."""
-        os.chdir(self.old_cwd)
-        shutil.rmtree(self.temp_dir)
 
 
 def _save_specification_in_temporary_file(
@@ -224,16 +138,24 @@ def _generate_protocol(package_path: Path) -> None:
     """
     cmd = ["generate", "protocol", os.path.join("..", package_path.name) + ".yaml"]
     log(f"Generate the protocol. Command: {pprint.pformat(cmd)}")
-    run_aea(*cmd)
+    AEAProject.run_aea(*cmd)
 
 
 def run_isort_and_black(directory: Path, **kwargs: Any) -> None:
     """Run black and isort against a directory."""
     kwargs["stdout"] = subprocess.PIPE
-    run_cli(
+
+    try:
+        # check if black and isort is installed
+        import black  # type: ignore  # noqa  # pylint: disable=unused-import,import-outside-toplevel
+        import isort  # type: ignore  # noqa  # pylint: disable=unused-import,import-outside-toplevel
+    except ImportError as e:
+        raise click.ClickException(str(e)) from e
+
+    AEAProject.run_cli(
         sys.executable, "-m", "black", "-q", str(directory.absolute()), **kwargs,
     )
-    run_cli(
+    AEAProject.run_cli(
         sys.executable,
         "-m",
         "isort",
@@ -242,23 +164,6 @@ def run_isort_and_black(directory: Path, **kwargs: Any) -> None:
         str(directory.absolute()),
         **kwargs,
     )
-
-
-def replace_in_directory(name: str, replacement_pairs: List[Tuple[str, str]]) -> None:
-    """
-    Replace text in directory.
-
-    :param name: the protocol name.
-    :param replacement_pairs: a list of pairs of strings (to_replace, replacement).
-    """
-    log(f"Replace prefix of import statements in directory '{name}'")
-    package_dir = Path(PROTOCOLS_PLURALS, name)
-    for submodule in package_dir.rglob("*.py"):
-        log(f"Process submodule {submodule.relative_to(package_dir)}")
-        for to_replace, replacement in replacement_pairs:
-            if to_replace not in submodule.read_text():
-                continue
-            submodule.write_text(submodule.read_text().replace(to_replace, replacement))
 
 
 def _fix_generated_protocol(package_path: Path) -> None:
@@ -275,17 +180,15 @@ def _fix_generated_protocol(package_path: Path) -> None:
     log(f"Restore original custom types in {package_path}")
     custom_types_module = package_path / CUSTOM_TYPE_MODULE_NAME
     if custom_types_module.exists():
-        file_to_replace = Path(
-            PROTOCOLS_PLURALS, package_path.name, CUSTOM_TYPE_MODULE_NAME
-        )
+        file_to_replace = Path(PROTOCOLS, package_path.name, CUSTOM_TYPE_MODULE_NAME)
         file_to_replace.write_text(custom_types_module.read_text())
 
-    package_readme_file = package_path / README_FILENAME
+    package_readme_file = package_path / DEFAULT_README_FILE
     if package_readme_file.exists():
         log(f"Copy the README {package_readme_file} into the new generated protocol.")
         shutil.copyfile(
             package_readme_file,
-            Path(PROTOCOLS_PLURALS, package_path.name, README_FILENAME),
+            Path(PROTOCOLS, package_path.name, DEFAULT_README_FILE),
         )
 
 
@@ -297,18 +200,16 @@ def _update_original_protocol(package_path: Path) -> None:
     """
     log(f"Copy the new protocol into the original directory {package_path}")
     shutil.rmtree(package_path)
-    shutil.copytree(Path(PROTOCOLS_PLURALS, package_path.name), package_path)
+    shutil.copytree(Path(PROTOCOLS, package_path.name), package_path)
 
 
 def _fingerprint_protocol(name: str) -> None:
     """Fingerprint the generated (and modified) protocol."""
     log(f"Fingerprint the generated (and modified) protocol '{name}'")
     protocol_config = load_component_configuration(
-        ComponentType.PROTOCOL,
-        Path(PROTOCOLS_PLURALS, name),
-        skip_consistency_check=True,
+        ComponentType.PROTOCOL, Path(PROTOCOLS, name), skip_consistency_check=True,
     )
-    run_aea("fingerprint", "protocol", str(protocol_config.public_id))
+    AEAProject.run_aea("fingerprint", "protocol", str(protocol_config.public_id))
 
 
 def _parse_generator_docstring(package_path: Path) -> str:
@@ -342,58 +243,10 @@ def _replace_generator_docstring(package_path: Path, replacement: str) -> None:
     :param replacement: the replacement to use.
     """
     protocol_name = package_path.name
-    init_module = Path(PROTOCOLS_PLURALS) / protocol_name / "__init__.py"
+    init_module = Path(PROTOCOLS) / protocol_name / "__init__.py"
     content = init_module.read_text()
     content = re.sub(PROTOCOL_GENERATOR_DOCSTRING_REGEX, replacement, content)
     init_module.write_text(content)
-
-
-def _process_packages_protocol(
-    package_path: Path,
-    preserve_generator_docstring: bool,
-    no_format: bool,
-    root_dir: Path,
-) -> None:
-    """
-    Process protocol package from local registry.
-
-    If the flag '--no-bump' is specified, it means the protocol generator
-    string that records the AEA and the protoc version used, i.e.:
-
-        It was created with protocol buffer compiler version `libprotoc ...` and aea version `...`.
-
-    must not be updated, as the 'AEA' version could have been changed.
-
-    It means:
-    - extract protocol specification from README
-    - generate the protocol in the current AEA project
-    - fix the generated protocol (e.g. import prefixed, custom types, ...)
-    - update the original protocol with the newly generated one.
-
-    It assumes the working directory is an AEA project.
-
-    :param package_path: path to the package.
-    :param preserve_generator_docstring: if True, the protocol generator docstring is preserved (see above).
-    :param no_format: If True, won't run the formatters
-    :param root_dir: Path to open-aea repository
-    """
-    if preserve_generator_docstring:
-        # save the old protocol generator docstring
-        old_protocol_generator_docstring = _parse_generator_docstring(package_path)
-    specification_content = get_protocol_specification_from_readme(package_path)
-    _save_specification_in_temporary_file(package_path.name, specification_content)
-    _generate_protocol(package_path)
-    _fix_generated_protocol(package_path)
-    if preserve_generator_docstring:
-        _replace_generator_docstring(package_path, old_protocol_generator_docstring)
-
-    if not no_format:
-        run_isort_and_black(
-            Path(PROTOCOLS_PLURALS, package_path.name), cwd=str(root_dir)
-        )
-
-    _fingerprint_protocol(package_path.name)
-    _update_original_protocol(package_path)
 
 
 def _check_preliminaries() -> None:
@@ -413,36 +266,6 @@ def _check_preliminaries() -> None:
         LIBPROTOC_VERSION in result_str,
         f"Invalid version for protoc. Found: {result_str}. Required: {LIBPROTOC_VERSION}.",
     )
-
-
-def _process_test_protocol(
-    specification: Path, package_path: Path, no_format: bool, root_dir: Path
-) -> None:
-    """
-    Process a test protocol.
-
-    :param specification: path to specification.
-    :param package_path: the output directory.
-    :param no_format: If True, won't run the formatters
-    :param root_dir: Path to open-aea repository
-    """
-    specification_content = specification.read_text()
-    _save_specification_in_temporary_file(package_path.name, specification_content)
-    _generate_protocol(package_path)
-    _fix_generated_protocol(package_path)
-    replacements = [
-        (
-            f"from packages.fetchai.protocols.{package_path.name}",
-            f"from tests.data.generator.{package_path.name}",
-        )
-    ]
-    replace_in_directory(package_path.name, replacements)
-    if not no_format:
-        run_isort_and_black(
-            Path(PROTOCOLS_PLURALS, package_path.name), cwd=str(root_dir)
-        )
-    _fingerprint_protocol(package_path.name)
-    _update_original_protocol(package_path)
 
 
 def download_package(package_id: PackageId, destination_path: str) -> None:
@@ -524,6 +347,53 @@ def _bump_protocol_specification_id_if_needed(package_path: Path) -> None:
     )
 
 
+def _process_packages_protocol(
+    package_path: Path,
+    preserve_generator_docstring: bool,
+    no_format: bool,
+    root_dir: Path,
+) -> None:
+    """
+    Process protocol package from local registry.
+
+    If the flag '--no-bump' is specified, it means the protocol generator
+    string that records the AEA and the protoc version used, i.e.:
+
+        It was created with protocol buffer compiler version `libprotoc ...` and aea version `...`.
+
+    must not be updated, as the 'AEA' version could have been changed.
+
+    It means:
+    - extract protocol specification from README
+    - generate the protocol in the current AEA project
+    - fix the generated protocol (e.g. import prefixed, custom types, ...)
+    - update the original protocol with the newly generated one.
+
+    It assumes the working directory is an AEA project.
+
+    :param package_path: path to the package.
+    :param preserve_generator_docstring: if True, the protocol generator docstring is preserved (see above).
+    :param no_format: If True, won't run the formatters
+    :param root_dir: Path to open-aea repository
+    """
+    if preserve_generator_docstring:
+        # save the old protocol generator docstring
+        old_protocol_generator_docstring = _parse_generator_docstring(package_path)
+    specification_content = get_protocol_specification_from_readme(package_path)
+    _save_specification_in_temporary_file(package_path.name, specification_content)
+    _generate_protocol(package_path)
+    _fix_generated_protocol(package_path)
+
+    if preserve_generator_docstring:
+        _replace_generator_docstring(package_path, old_protocol_generator_docstring)
+
+    if not no_format:
+        run_isort_and_black(Path(PROTOCOLS, package_path.name), cwd=str(root_dir))
+
+    _fingerprint_protocol(package_path.name)
+    _update_original_protocol(package_path)
+
+
 @click.command(name="generate-all-protocols")
 @click.argument(
     "packages_dir",
@@ -535,11 +405,6 @@ def _bump_protocol_specification_id_if_needed(package_path: Path) -> None:
     is_flag=True,
     default=False,
     help="no_bump implies to ignore the docstring",
-)
-@click.option(
-    "--test-data",
-    type=click.Path(dir_okay=True, file_okay=False),
-    help="Path to test data.",
 )
 @click.option(
     "--no-format",
@@ -559,7 +424,6 @@ def _bump_protocol_specification_id_if_needed(package_path: Path) -> None:
 def generate_all_protocols(
     packages_dir: Path,
     no_bump: bool,
-    test_data: Optional[Path],
     no_format: bool,
     root_dir: Path,
     check_clean: bool,
@@ -568,27 +432,18 @@ def generate_all_protocols(
 
     _check_preliminaries()
 
-    all_protocols = list(find_protocols_in_local_registry(packages_dir))
-    if test_data is not None:
-        test_data = Path(test_data).absolute()
+    packages_dir = Path(packages_dir)
+    all_protocols = find_protocols_in_local_registry(packages_dir)
 
     with AEAProject():
-        if test_data is not None:
-            for file, output_dir in DEFAULT_TEST_SAMPLES:
-                log_separator()
-                _process_test_protocol(
-                    test_data / file,
-                    test_data / "generator" / output_dir,
-                    no_format,
-                    root_dir,
-                )
-
         for package_path in all_protocols:
             log_separator()
             log(f"Processing protocol at path {package_path}")
             if not no_bump:
                 _bump_protocol_specification_id_if_needed(package_path)
-            _process_packages_protocol(package_path, no_bump, no_format, root_dir)
+            _process_packages_protocol(
+                package_path, no_bump, no_format, root_dir,
+            )
 
     if check_clean:
         check_working_tree_is_dirty()
